Initial support for the SolidRun CN913x EVK boards
This is the initial conversion of the runme.sh script and patches from
SolidRun, https://github.com/SolidRun/cn913x_build GIT hash cead42e32.

The build is based on the CN9130 CRB config, but using the version of
U-Boot that SolidRun patches (older then the CRB).

Current state:

  - The board boots and gets passed U-Boot
  - Finds the extlinux.conf and presents the boot menu
  - Loads the correct device tree (see patches)
  - Loads the Linux Image
  - Gets stuck

TODO:

  - Enable early console to debug
  - Investigate https://github.com/SolidRun/cn913x_build/pull/1 to
    see if the updates+changes proposed there helps

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>